Sett up set ang get OSC routes for seed and direction

+++ Teensy/lib/PinneRobot/DCMotorDriver.cpp
@@ -4,9 +4,7 @@ speed_t DCMotorDriver::GetSpeed() {
 	return _speed;
 }
 
-int DCMotorDriver::GetDirection() {
-	return _direction;
-}
+uint8_t DCMotorDriver::GetDirection() { return _direction; }
 
 speed_t DCMotorDriver::GetBrake() {
 	return _brake;
